Remove the From: header from original mail
Apparently, Postfix sends the mail with the original From: header, which isn't
rewritten, causing DKIM and DMARC to fail validation.

This commit removes the From: header, and adds the -F parameter to sendmail to
specify a nicer sender name. When the From: header was removed, the sender
name is set to "nobody", which is ugly.